backend/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
sqlalchemy[asyncio]>=2.0,<3
asyncmy>=0.2.9,<1
python-dotenv>=1.0,<2
pydantic-settings>=2.2,<3
pyjwt>=2.8,<3
bcrypt>=4.1,<5
